Revert "refactor: add context to own module"
This reverts commit 2ed6aa1da922b80b64e50f9bdcef1af14fbaeb29.
