+++ src/user/github-profile.service.ts
@@ -15,7 +15,7 @@ export class GithubProfileService {
     private readonly githubModel: Model<GithubProfile>,
     private readonly geocodingService: GeocodingService,
     private readonly httpService: HttpService,
-  ) {}
+  ) { }
 
   public async create(username: string) {
     const location = await lastValueFrom(
@@ -27,12 +27,7 @@ export class GithubProfileService {
             }
           });
           if (!githubData.location) return { ...githubData };
-          return {
-            ...githubData,
-            location: await this.geocodingService.fetchCoordinates(
-              githubData.location,
-            ),
-          };
+          return { ...githubData, location: await this.geocodingService.fetchCoordinates(githubData.location) };
         }),
       ),
     );
@@ -51,12 +46,10 @@ export class GithubProfileService {
   public async delete(username: string) {
     try {
       return await this.githubModel.findByIdAndDelete(username);
-    } catch (error) {
+    }
+    catch (error) {
       console.log(error);
-      throw new HttpException(
-        `Github-Profile for ${username} could not be deleted`,
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
+      throw new HttpException(`Github-Profile for ${username} could not be deleted`, HttpStatus.INTERNAL_SERVER_ERROR)
     }
   }
 
@@ -71,13 +64,15 @@ export class GithubProfileService {
           );
         }),
         map((response) => response.data),
-        map((githubProfileResponse: GithubProfileResponse) => {
-          return {
-            location: githubProfileResponse.location,
-            repos: githubProfileResponse.public_repos,
-            followers: githubProfileResponse.followers,
-          };
-        }),
+        map(
+          (githubProfileResponse: GithubProfileResponse) => {
+            return {
+              location: githubProfileResponse.location,
+              repos: githubProfileResponse.public_repos,
+              followers: githubProfileResponse.followers
+            }
+          }
+        ),
       );
   }
 }
